Package Firefox debug adapter
+ some refactorings
