+++ org.eclipse.wildwebdeveloper/src/org/eclipse/wildwebdeveloper/debug/node/NodeRunDAPDebugDelegate.java
@@ -11,11 +11,12 @@
  *   Mickael Istria (Red Hat Inc.) - initial implementation
  *   Pierre-Yves B. - Issue #180 Wrong path to nodeDebug.js
  *******************************************************************************/
-package org.eclipse.wildwebdeveloper.debug;
+package org.eclipse.wildwebdeveloper.debug.node;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -26,43 +27,69 @@ import org.eclipse.core.runtime.FileLocator;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
+import org.eclipse.debug.core.DebugPlugin;
 import org.eclipse.debug.core.ILaunch;
 import org.eclipse.debug.core.ILaunchConfiguration;
+import org.eclipse.debug.core.ILaunchManager;
 import org.eclipse.jface.dialogs.ErrorDialog;
+import org.eclipse.lsp4e.debug.DSPPlugin;
 import org.eclipse.lsp4e.debug.launcher.DSPLaunchDelegate;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.wildwebdeveloper.Activator;
 import org.eclipse.wildwebdeveloper.InitializeLaunchConfigurations;
 
-public class FirefoxAttachDebugDelegate extends DSPLaunchDelegate {
+import com.google.gson.JsonObject;
+
+public class NodeRunDAPDebugDelegate extends DSPLaunchDelegate {
 
 	static final String ID = "org.eclipse.wildwebdeveloper.launchConfiguration.nodeDebug"; //$NON-NLS-1$
 
-	// see https://github.com/firefox-devtools/vscode-firefox-debug/blob/master/src/adapter/configuration.ts for launch/attach configuration parameters
-	static final String PORT = "port"; //$NON-NLS-1$
-	static final String REQUEST = "request"; //$NON-NLS-1$
-
+	// see https://github.com/Microsoft/vscode-node-debug/blob/master/src/node/nodeDebug.ts LaunchRequestArguments
+	static final String PROGRAM = "program"; //$NON-NLS-1$
+	public static final String ARGUMENTS = "args"; //$NON-NLS-1$
+	private static final String CWD = "cwd"; //$NON-NLS-1$
+	private static final String ENV = "env"; //$NON-NLS-1$
 
 	@Override
 	public void launch(ILaunchConfiguration configuration, String mode, ILaunch launch, IProgressMonitor monitor)
 			throws CoreException {
 		// user settings
-		Map<String, Object> param = new HashMap<>(); 
-		param.put(REQUEST, "attach"); //$NON-NLS-1$
-		int port = configuration.getAttribute(NodeAttachDebugDelegate.PORT, 4711);
-		param.put(PORT, port);
+		Map<String, Object> param = new HashMap<>();
+		param.put(PROGRAM, configuration.getAttribute(PROGRAM, "no program path defined")); //$NON-NLS-1$
+		String argsString = configuration.getAttribute(ARGUMENTS, "").trim(); //$NON-NLS-1$
+		if (!argsString.isEmpty()) {
+			Object[] args = Arrays.asList(argsString.split(" ")).stream() //$NON-NLS-1$
+					.filter(s -> !s.trim().isEmpty()).toArray();
+			if (args.length > 0) {
+				param.put(ARGUMENTS, args);
+			}
+		}
+		Map<String, String> env = configuration.getAttribute(ILaunchManager.ATTR_ENVIRONMENT_VARIABLES,
+				Collections.emptyMap());
+		if (!env.isEmpty()) {
+			JsonObject envJson = new JsonObject();
+			env.forEach((key, value) -> envJson.addProperty(key, value));
+			param.put(ENV, envJson);
+		}
+		String cwd = configuration.getAttribute(DebugPlugin.ATTR_WORKING_DIRECTORY, "").trim(); //$NON-NLS-1$
+		if (!cwd.isEmpty()) {
+			param.put(CWD, cwd);
+		}
+		// workaround until
+		// https://github.com/microsoft/vscode-node-debug2/commit/f2dfa4ca4026fb3e4f143a391270a03df8187b42#diff-d03a74f75ec189cbc7dd3d2e105fc9c9R625
+		// is released in VSCode
+		param.put("sourceMaps", false);
 
 		try {
-			// TODO: Properly package Firefox Debug Adapter
 			URL fileURL = FileLocator.toFileURL(
-					getClass().getResource("/org.eclipse.wildwebdeveloper/language-servers/node_modules/vscode-firefox-debug/out/adapter/firefoxDebugAdapter.js"));
-			File file = new File("/home/aobuchow/git/vscode-firefox-debug/out/adapter/firefoxDebugAdapter.js");
+					getClass().getResource("/language-servers/node_modules/node-debug2/out/src/nodeDebug.js"));
+			File file = new File(fileURL.getPath());
 			List<String> debugCmdArgs = Collections.singletonList(file.getAbsolutePath());
 
 			DSPLaunchDelegateLaunchBuilder builder = new DSPLaunchDelegateLaunchBuilder(configuration, mode, launch,
 					monitor);
 			builder.setLaunchDebugAdapter(InitializeLaunchConfigurations.getNodeJsLocation(), debugCmdArgs);
-			builder.setMonitorDebugAdapter(true);
+			builder.setMonitorDebugAdapter(configuration.getAttribute(DSPPlugin.ATTR_DSP_MONITOR_DEBUG_ADAPTER, false));
 			builder.setDspParameters(param);
 
 			super.launch(builder);
